chat be done

+++ controllers/messageControllers.js
@@ -1,3 +1,4 @@
+import { validationResult } from "express-validator";
 import { messages } from "../models/messages.js";
 import { catchAsync } from "../utils/catchAsync.js";
 
@@ -12,16 +13,23 @@ export const sendMessage = catchAsync(async (req, res, next) => {
     const { matchId, message } = req.body
     const sender = req.user._id
 
-    const data = await messages.create({ message, matchId, reciever, sender });
-
+    const data = await messages.create({ message, matchId, reciever, sender })
+    await data.populate('sender')
+    console.log(data)
     return res.status(200).json({
         message:"message sent succesfully",
         data:data
     })
 })
 export const getMessages = catchAsync(async(req, res, next) => {
+    const result = validationResult(req);
+
+    if (!result.isEmpty()) {
+        return res.status(400).json(result)
+    }
+    console.log(req.query)
     const {matchId} = req.query;
-    const data = await messages.find({matchId})
+    const data = await messages.find({matchId}).populate('sender')
     return res.status(200).json({
         message:"messages fetched successfully",
         data:data
@@ -30,6 +38,12 @@ export const getMessages = catchAsync(async(req, res, next) => {
 
 
 export const editMessage = catchAsync(async(req, res, next) => {
+    const result = validationResult(req);
+
+    if (!result.isEmpty()) {
+        return res.status(400).json(result)
+    }
+    console.log(req.query)
     const {message} = req.body;
     const {matchId} = req.query;
 
@@ -44,6 +58,14 @@ export const editMessage = catchAsync(async(req, res, next) => {
 
 
 export const deleteMessage = catchAsync(async(req, res, next) => {
+    const result = validationResult(req);
+
+    if (!result.isEmpty()) {
+        return res.status(400).json(result)
+    }
+    
+    console.log(req.query)
+
     const {messageId} = req.query;
     await messages.findByIdAndDelete(messageId);
     return res.status(200).json({
